backend/requirements.txt
alembic==1.14.0
babel==2.17.0
bcrypt==4.2.1
certifi==2024.8.30
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
comm==0.2.2
cryptography==45.0.3
css==0.1
decorator==5.1.1
docx==0.2.4
docx2txt==0.8
docxcompose==1.4.0
docxtpl==0.20.0
executing==2.1.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.10
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
gevent==24.10.3
google==3.0.0
gunicorn==23.0.0
h11==0.14.0
html2docx==1.6.0
idna==3.10
ipykernel==6.29.5
jsonpath-python==1.0.6
mysql-connector-python==9.3.0
networkx==3.4.2
openai==1.55.0
PyJWT==2.10.1
PyMySQL==1.1.1
PyPDF2==3.0.1
python-docx
python-dotenv
python-magic
python-multipart==0.0.12
python-pptx==0.6.18
pytz==2024.2
regex==2024.11.6
requests==2.32.3
rich==13.9.3
simplejson==3.20.1
six==1.12.0
SQLAlchemy==2.0.36
sympy==1.13.1
textract==1.6.3
tiktoken==0.8.0
tokenizers==0.20.2
torch==2.5.1
typer==0.12.5
urllib3==2.2.3
websocket==0.2.1
websockets==12.0
Werkzeug==3.1.3